Add gfx_device::GfxDevice::render_world() method
* Divide Frame into Scene and Pipeline
* Move loading code and Texture and Mesh into
  gfx_device::texture/gfx_device::mesh modules
* Comment out rendering processor
