+++ Cargo.toml
@@ -60,3 +60,8 @@ gfx = "0.12"
 cgmath="0.11"
 
 [workspace]
+glutin = "0.6"
+gfx = "0.12"
+gfx_device_gl = "0.11"
+gfx_window_glutin = "0.12"
+genmesh="0.4"
