+++ examples/01_window/resources/config.yml
@@ -1,11 +1,11 @@
----
-display_config:
-    dimensions: null
-    fullscreen: false 
-    max_dimensions: null
-    min_dimensions: null
-    multisampling: 1
-    title: "Windowing example"
-    visibility: true
-    vsync: true
-    backend: OpenGL
+# ---
+# display_config:
+dimensions: null
+fullscreen: false 
+max_dimensions: null
+min_dimensions: null
+multisampling: 1
+title: "Windowing example"
+visibility: true
+vsync: true
+backend: OpenGL
