+++ src/asset_manager/mod.rs
@@ -1,0 +1,2 @@
+mod asset_manager;
+pub use self::asset_manager::*;
