+++ src/engine/mod.rs
@@ -2,8 +2,8 @@
 
 mod app;
 mod state;
-mod config;
+// mod config;
 
 pub use self::app::{Application, ApplicationBuilder};
 pub use self::state::{State, StateMachine, Trans};
-pub use self::config::Config;
+// pub use self::config::Config;
